Made saves work properlly on 64bit linux
changed fileio to use unit*_t types

+++ series.c
@@ -452,7 +452,7 @@ static int getseriesfile(char *filename, void *data)
     fileinfo		file;
     seriesdata	       *sdata = (seriesdata*)data;
     gameseries	       *series;
-    unsigned long	magic;
+    uint32_t		magic;
     char	       *datfilename;
     int			config, f;
